requirements.txt
acres==0.5.0
certifi==2025.7.14
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.18.0
hidapi==0.14.0.post4
httplib2==0.22.0
idna==3.10
keyboard==0.13.5
looseversion==1.3.0
lxml==6.0.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
prov==2.1.1
puremagic==1.30
pydot==4.0.1
PyMuPDF==1.26.3
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
pytuyo @ git+https://github.com/rabryan/pytuyo.git@251848e96f8d331f11012f94e3fbe771e2fd34e5
pytz==2025.2
pyusb==1.3.1
pyxnat==1.6.3
rdflib==7.1.4
requests==2.32.4
scipy==1.16.1
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
traits==7.0.2
tzdata==2025.2
urllib3==2.5.0
